[libc++] Partial reversion of 21f73d5826fb50 (using_if_exists).
The premise of the patch is that either libc++ is used with a compiler
supporting attribute using_if_exists, or the underlying OS will support
all the interfaces so guarded.

Neither of these is true when using GCC on Apple platforms, so that we
need to re-introduce handling for the case that we are on macOS and the
compiler does not support attribute using_if_exists.